# kxmj.core.api/docs/swagger.yaml
basePath: /api/v1
definitions:
...
  model.AndroidInfo:
    properties:
      id:
        description: android id
        type: string
      imei:
        description: 自定义imei
        type: string
      sdk:
        description: android sdk 版本
        type: integer
    type: object
  model.BindPhoneNumReq:
    properties:
      SMSCode:
        description: 短信验证码
        type: string
      telNumber:
        description: 手机号码
        type: string
      userId:
        description: user id
        type: integer
    required:
    - SMSCode
    - telNumber
    - userId
    type: object
  model.BindPhoneNumResp:
    properties:
      Message:
        description: 绑定成功通知
        type: string
    type: object
...
  model.GetAppBaseInfoResp:
    properties:
      hotRenewAddress:
        description: 热更新包地址
        type: string
      wechatAppId:
        description: 微信开放平台应用唯一标识
        type: string
      wechatSecretKey:
        description: 微信登陆应用密钥APPSecret
        type: string
    type: object
...
  model.GetGatewayInfo:
    properties:
      addr:
        type: string
      port:
        type: integer
      svrId:
        type: integer
      svrType:
        type: integer
    type: object
  model.GetGatewayResp:
    properties:
      list:
        items:
          $ref: '#/definitions/model.GetGatewayInfo'
        type: array
    type: object
  model.GetHomeResp:
    properties:
      guides:
        additionalProperties:
          type: integer
        description: 首页引导提示列表(key：类型：1 商城；2 背包；3 活动；4 福利；5 直播；6 签到任务；7 对局任务；8 赢金任务；9
          充值任务； value：提示数量)
        type: object
      user:
        $ref: '#/definitions/model.HomeUser'
        description: 用户首页信息
    type: object
  model.GetInfoResp:
    properties:
      avatarAddr:
        description: 头像地址
        type: string
      avatarFrame:
        description: 头像框
        type: integer
      diamond:
        description: 钻石数
        type: string
      gold:
        description: 金币数
        type: string
      goldBean:
        description: 金豆数
        type: string
      nickname:
        description: 昵称
        type: string
      userId:
        description: UID
        type: integer
      vip:
        description: VIP等级
        type: integer
    type: object
...
  model.HomeUser:
    properties:
      avatarAddr:
        description: 头像地址
        type: string
      avatarFrame:
        description: 头像框
        type: integer
      diamond:
        description: 钻石数
        type: string
      gender:
        description: 性别：0 女；1 男；
        type: integer
      gold:
        description: 金币数
        type: string
      goldBean:
        description: 金豆数
        type: string
      nickname:
        description: 昵称
        type: string
    type: object
...
  model.SyncDeviceReq:
    properties:
      androidInfo:
        $ref: '#/definitions/model.AndroidInfo'
        description: android 设备信息
      brand:
        description: 品牌
        type: string
      deviceId:
        description: 设备ID
        type: string
      height:
        description: 高度
        type: integer
      iosUUID:
        description: IOS设备ID
        type: string
      manufacturer:
        description: 制造商
        type: string
      model:
        description: 型号
        type: string
      organic:
        description: organic 1 自然、2 广告
        type: integer
      os:
        description: '系统: 0 未知 1 安卓 2 IOS 3 其它'
        type: integer
      version:
        description: 系统版本号
        type: string
      width:
        description: 宽度
        type: integer
    required:
    - deviceId
    type: object
  model.SyncDeviceResp:
    properties:
      deviceId:
        description: 设备Id
        type: string
    type: object
...
  model.TelRegisterReq:
    properties:
      SMSCode:
        description: 短信验证码
        type: string
      password:
        description: 密码
        type: string
      telNumber:
        description: 手机号码
        type: string
    required:
    - SMSCode
    - password
    - telNumber
    type: object
  model.TelRegisterResp:
    properties:
      userId:
        description: user id
        type: integer
    type: object
...
  web.Response:
    properties:
      code:
        type: integer
      data: {}
      msg:
        type: string
    type: object
info:
  contact: {}
  description: core-api 服务接口
  title: core-api 服务接口
  version: "1.0"
paths:
  /app/get-base:
    get:
      description: APP
      parameters:
      - description: dev.kxmj.com
        in: header
        name: BundleId
        required: true
        type: string
      - description: 0033-0000-9999-9999-9999-1111
        in: header
        name: DeviceId
        required: true
        type: string
      - description: 9cb9aaf4e28d5094fbca383255ecbafbbabb95c40c758843054480391f448ee8
        in: header
        name: Authorization
        required: true
        type: string
      responses:
        "200":
          description: 请求成功
          schema:
            allOf:
            - $ref: '#/definitions/web.Response'
            - properties:
                data:
                  $ref: '#/definitions/model.GetAppBaseInfoResp'
              type: object
      summary: 获取APP基础信息
      tags:
      - APP
  /app/get-gateways:
    get:
      description: APP
      parameters:
      - description: dev.kxmj.com
        in: header
        name: BundleId
        required: true
        type: string
      - description: 0033-0000-9999-9999-9999-1111
        in: header
        name: DeviceId
        required: true
        type: string
      - description: 9cb9aaf4e28d5094fbca383255ecbafbbabb95c40c758843054480391f448ee8
        in: header
        name: Authorization
        required: true
        type: string
      responses:
        "200":
          description: 请求成功
          schema:
            allOf:
            - $ref: '#/definitions/web.Response'
            - properties:
                data:
                  $ref: '#/definitions/model.GetGatewayResp'
              type: object
      summary: 获取网关地址
      tags:
      - APP
  /app/get-home:
    get:
      description: APP
      parameters:
      - description: dev.kxmj.com
        in: header
        name: BundleId
        required: true
        type: string
      - description: 0033-0000-9999-9999-9999-1111
        in: header
        name: DeviceId
        required: true
        type: string
      - description: 9cb9aaf4e28d5094fbca383255ecbafbbabb95c40c758843054480391f448ee8
        in: header
        name: Authorization
        required: true
        type: string
      responses:
        "200":
          description: 请求成功
          schema:
            allOf:
            - $ref: '#/definitions/web.Response'
            - properties:
                data:
                  $ref: '#/definitions/model.GetHomeResp'
              type: object
      summary: 获取首页信息
      tags:
      - APP
...
  /app/sync-device:
    post:
      description: APP
      parameters:
      - description: dev.kxmj.com
        in: header
        name: BundleId
        required: true
        type: string
      - description: JSON
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/model.SyncDeviceReq'
      responses:
        "200":
          description: 请求成功
          schema:
            allOf:
            - $ref: '#/definitions/web.Response'
            - properties:
                data:
                  $ref: '#/definitions/model.SyncDeviceResp'
              type: object
      summary: 同步设备信息
      tags:
      - APP
...
  /user/bind-phone:
    post:
      description: USER
      parameters:
      - description: dev.kxmj.com
        in: header
        name: BundleId
        required: true
        type: string
      - description: 0033-0000-9999-9999-9999-1111
        in: header
        name: DeviceId
        required: true
        type: string
      - description: JSON
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/model.BindPhoneNumReq'
      responses:
        "200":
          description: 请求成功
          schema:
            allOf:
            - $ref: '#/definitions/web.Response'
            - properties:
                data:
                  $ref: '#/definitions/model.BindPhoneNumResp'
              type: object
      summary: 绑定手机号
      tags:
      - USER
...
  /user/get-info:
    get:
      description: USER
      parameters:
      - description: dev.kxmj.com
        in: header
        name: BundleId
        required: true
        type: string
      - description: 0033-0000-9999-9999-9999-1111
        in: header
        name: DeviceId
        required: true
        type: string
      - description: 9cb9aaf4e28d5094fbca383255ecbafbbabb95c40c758843054480391f448ee8
        in: header
        name: Authorization
        required: true
        type: string
      responses:
        "200":
          description: 请求成功
          schema:
            allOf:
            - $ref: '#/definitions/web.Response'
            - properties:
                data:
                  $ref: '#/definitions/model.GetInfoResp'
              type: object
      summary: 获取用户信息
      tags:
      - USER
...
  /user/tel-register:
    post:
      description: USER
      parameters:
      - description: dev.kxmj.com
        in: header
        name: BundleId
        required: true
        type: string
      - description: 0033-0000-9999-9999-9999-1111
        in: header
        name: DeviceId
        required: true
        type: string
      - description: JSON
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/model.TelRegisterReq'
      responses:
        "200":
          description: 请求成功
          schema:
            allOf:
            - $ref: '#/definitions/web.Response'
            - properties:
                data:
                  $ref: '#/definitions/model.TelRegisterResp'
              type: object
      summary: 手机注册
      tags:
      - USER
...
schemes:
- http
- https
securityDefinitions:
  ApiKeyAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"